LSTM_ArrayNet_code/requirements.txt
torch>=2.0.0
numpy>=1.24.0
scipy>=1.10.1  # 用于读取MATLAB的.mat文件
scikit-learn>=1.2.2
PyYAML>=6.0
matplotlib>=3.7.1
tensorboardX>=2.6.2  # 用于训练可视化